+++ src/components/subs/TranslateAlternativesItem.tsx
@@ -18,27 +18,47 @@ function TranslateAlternativesItem(props: Props) {
 
   function handleOnClick(e: any) {
     e.stopPropagation();
-    const partOfSpeechNode = document.querySelector(`#part-of-speech-${props.groupIndex}`);
+    const partOfSpeechNode = document.querySelector(
+      `#part-of-speech-${props.groupIndex}`
+    );
 
     props.currentService
-      .addWord(Utils.clearWord(props.word), translateNode.current.textContent, partOfSpeechNode.textContent)
+      .addWord(
+        Utils.clearWord(props.word),
+        translateNode.current.textContent,
+        partOfSpeechNode.textContent
+      )
       .then((text: string) => (toast as any).info(text))
       .catch((error: string) => (toast as any).error(error));
   }
 
   return [
-    props.currentService
-      ? <td key={"plus"} className="easysubs-translate-alternative-item-add-to-learn" onClick={handleOnClick}>
-          <Plus style={{ fill: props.currentService.color }} />
-        </td>
-      : null,
-    <td key={"translate"} className="easysubs-translate-alternative-item-translate" ref={translateNode}>
+    props.currentService ? (
+      <td
+        key={"plus"}
+        className="easysubs-translate-alternative-item-add-to-learn"
+        onClick={handleOnClick}
+      >
+        <Plus style={{ fill: props.currentService.color }} />
+      </td>
+    ) : null,
+    <td
+      key={"translate"}
+      className="easysubs-translate-alternative-item-translate"
+      ref={translateNode}
+    >
       {props.alternative[0]}
     </td>,
-    <td key={"original"} className="easysubs-translate-alternative-item-original">
-      {props.alternative[1].slice(0, 3).join(", ")}
+    <td
+      key={"original"}
+      className="easysubs-translate-alternative-item-original"
+    >
+      {props.alternative[1]?.slice(0, 3)?.join(", ")}
     </td>,
-    <td key={"frequency"} className="easysubs-translate-alternative-item-frequency">
+    <td
+      key={"frequency"}
+      className="easysubs-translate-alternative-item-frequency"
+    >
       <FrequencyDots frequency={props.alternative[3]} />
     </td>
   ] as any;